Update for contract changes (uint128 > uint256)

+++ contracts/CoinFlip.sol
@@ -6,7 +6,7 @@
 pragma solidity ^0.8.16;
 
 interface IRandomizer {
-    function request(uint256 callbackGasLimit) external returns (uint128);
+    function request(uint256 callbackGasLimit) external returns (uint256);
 
     function clientWithdrawTo(address to, uint256 amount) external;
 }
@@ -23,7 +23,7 @@ contract CoinFlip {
 
     event FlipResult(
         address indexed player,
-        uint128 indexed id,
+        uint256 indexed id,
         uint256 seed,
         bool prediction,
         bool result
@@ -31,9 +31,9 @@ contract CoinFlip {
 
     event OwnerUpdated(address indexed user, address indexed newOwner);
 
-    mapping(uint128 => CoinFlipGame) coinFlipGames;
-    mapping(address => uint128[]) userToGames;
-    mapping(uint128 => bool) gameToHeadsTails;
+    mapping(uint256 => CoinFlipGame) coinFlipGames;
+    mapping(address => uint256[]) userToGames;
+    mapping(uint256 => bool) gameToHeadsTails;
 
     address public owner;
     address public proposedOwner;
@@ -49,14 +49,14 @@ contract CoinFlip {
     // Called by player to initiate a coinflip
     // Using randomizer's request id as the game id
     function flip(bool prediction) external {
-        uint128 id = IRandomizer(randomizer).request(20000);
+        uint256 id = IRandomizer(randomizer).request(20000);
         userToGames[msg.sender].push(id);
         coinFlipGames[id] = CoinFlipGame(msg.sender, prediction, false, 0);
         emit Flip(msg.sender, id, prediction);
     }
 
     // The callback function called by randomizer when the random bytes are ready
-    function randomizerCallback(uint128 _id, bytes32 _value) external {
+    function randomizerCallback(uint256 _id, bytes32 _value) external {
         require(
             msg.sender == address(randomizer),
             "Only the randomizer contract can call this function"
@@ -70,7 +70,7 @@ contract CoinFlip {
     }
 
     // Get a game struct for a game id
-    function getGame(uint128 _id) external view returns (CoinFlipGame memory) {
+    function getGame(uint256 _id) external view returns (CoinFlipGame memory) {
         return coinFlipGames[_id];
     }
 
@@ -78,7 +78,7 @@ contract CoinFlip {
     function getPlayerGameIds(address _player)
         external
         view
-        returns (uint128[] memory)
+        returns (uint256[] memory)
     {
         return userToGames[_player];
     }
